requirements.txt
fastapi>=0.110
httpx>=0.27
jinja2>=3.1
pydantic>=2.6
PyYAML>=6.0
python-dotenv>=1.0
uvicorn[standard]>=0.27
